requirements.txt
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0  # PostgreSQL 数据库支持（可选，用于持久化存储）